Initial work to detect methods marked with `[ParallelQueryAttribute]`
