+++ Arch.System.SourceGenerator/Extensions/GeneratorSyntaxContextExtensions.cs
@@ -1,0 +1,41 @@
+﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Arch.System.SourceGenerator.Extensions;
+
+internal static class GeneratorSyntaxContextExtensions
+{
+    /// <summary>
+    ///     Returns a <see cref="MethodDeclarationSyntax"/> if its annocated with a attribute of <see cref="name"/>.
+    /// </summary>
+    /// <param name="context">Its <see cref="GeneratorSyntaxContext"/>.</param>
+    /// <param name="name">The attributes name.</param>
+    /// <returns></returns>
+    public static MethodDeclarationSyntax? GetMethodSymbolIfAttributeOf(ref this GeneratorSyntaxContext context, string name)
+    {
+        // we know the node is a EnumDeclarationSyntax thanks to IsSyntaxTargetForGeneration
+        var enumDeclarationSyntax = (MethodDeclarationSyntax)context.Node;
+
+        // loop through all the attributes on the method
+        foreach (var attributeListSyntax in enumDeclarationSyntax.AttributeLists)
+        {
+            foreach (var attributeSyntax in attributeListSyntax.Attributes)
+            {
+                if (context.SemanticModel.GetSymbolInfo(attributeSyntax).Symbol is not IMethodSymbol attributeSymbol)
+                    continue;
+
+                var attributeContainingTypeSymbol = attributeSymbol.ContainingType;
+                var fullName = attributeContainingTypeSymbol.ToDisplayString();
+
+                // Is the attribute the [EnumExtensions] attribute?
+                if (fullName != name)
+                    continue;
+
+                return enumDeclarationSyntax;
+            }
+        }
+
+        // we didn't find the attribute we were looking for
+        return null;
+    }
+}
