+++ Arch.System/Attributes.cs
@@ -8,6 +8,15 @@ public class QueryAttribute : global::System.Attribute
 {
 }
 
+/// <summary>
+///     Marks a method to generate a high performance parallelised query for it. 
+/// </summary>
+[global::System.AttributeUsage(global::System.AttributeTargets.Method)]
+public class ParallelQueryAttribute
+    : global::System.Attribute
+{
+}
+
 
 /// <summary>
 ///     Marks a parameter as "data". This will be taken into account during source generation and will still be passed as a parameter in the query method.
